requirements.txt
docling==2.80.0
openai==2.29.0
requests==2.32.5
PyYAML==6.0.3

# Development and test dependency
pytest==9.0.2
